requirements.txt
pykeyboard
pretty-errors
faker
future
emoji
lxml
beautifulsoup4
requests
sqlalchemy==1.3.20
python-telegram-bot==12.8
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
googletrans==3.1.0a0
jikanpy
speedtest-cli
coffeehouse
regex
bleach
git+https://github.com/starry69/python-markdown2.git
wikipedia
telethon==1.16.4
spamwatch
alphabet_detector
pyrate-limiter
cachetools
pyrogram>=1.0.7
tgCrypto>=1.2.2
fontTools
gtts
bs4
apscheduler
heroku3
flag
telegraph
tswift
pytz
PyYAML>=5.3.1
redis
html2text
pythonping
html_telegraph_poster
pytube
youtube-search-python
instantmusic
PyLyrics
hachoir
pybase64
pySmartDL
validators
bing_image_downloader
google-api-python-client
wget
fake-useragent
asyncio
youtube_dl

glitch_this
NumPy
opencv-python-headless
envparse
loguru

cloudmersive_virus_api_client==3.0.1
pymongo==3.11.0
dnspython==2.0.0
pygments
tswift
youtube_search
fontTools
motor

pygithub
search_engine_parser

aiofiles
yt_dlp
youtube_search

#googletrans
google_trans_new

#new
odmantic==0.3.5
better_profanity
nudepy
secureme
python-arq
emoji-country-flag
countryinfo
PyDictionary==2.0.1

gpytranslate
pytz
